Allow for elements to be displayed in descending order
Up and down are relative. The assumption that the order of elements in the DOM will be congruent with the order of the sort is not correct when elements are displayed in descending order. With this change we now accommodate display of sortable elements in either ascending or descending order. (It's possible this is related to other issues with persistent reordering: https://github.com/RubaXa/Sortable/issues/398)